requirements.txt
django==5.0.7
requests==2.32.3
google-generativeai==0.7.2
python-dotenv==1.0.1
django-cors-headers==4.4.0
appwrite==5.0.1
firebase-admin==6.5.0
whitenoise==6.7.0
gunicorn==22.0.0
psycopg2-binary # PostgreSQL adapter
dj-database-url # For parsing DATABASE_URL